requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
pycrypto==2.6.1
pyOpenSSL==19.0.0
PySocks==1.6.8
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.2
